Carro API

+++ API/carros_api.py
@@ -1,6 +1,6 @@
 from flask import Flask, jsonify, request, Blueprint
 from psycopg2.extras import RealDictCursor
-from connect_db import connectToDb
+from connect_db import connectDb
 import psycopg2
 
 carros_api = Blueprint('carros_api', 'carros_api', url_prefix="/api/carros")
@@ -8,8 +8,8 @@ carros_api = Blueprint('carros_api', 'carros_api', url_prefix="/api/carros")
 @carros_api.route('/', methods=['GET', 'POST', 'PUT'])
 def api_verbs():
     if request.method == 'GET':
-        conn = connectToDb()
-        cur = conn.cursor(cursor_factory=RealDictCursor)
+        con = connectDb()
+        cur = con.cursor(cursor_factory=RealDictCursor)
 
         try:
             cur.execute('select * from tb_carros')
@@ -18,17 +18,19 @@ def api_verbs():
         except (Exception, psycopg2.DatabaseError) as error:
             print(error)
         finally:
-            if conn is not None:
-                conn.close()
+            if con is not None:
+                con.close()
 
+        if(result == None):
+            return 'Nenhum carro encontrado'
 
         return jsonify(result)
 
     elif request.method == 'POST':
         dados = request.json
 
-        conn = connectToDb()
-        cur = conn.cursor()
+        con = connectDb()
+        cur = con.cursor()
 
         try:
             cur.execute('''
@@ -40,23 +42,30 @@ def api_verbs():
                         dados['ano'],
                         dados['cor'],
                         dados['valor']))
-            conn.commit()
+            con.commit()
             cur.close()
         except (Exception, psycopg2.DatabaseError) as error:
             print(error)
         finally:
-            if conn is not None:
-                conn.close()
-
-        return "Carro inserido com sucesso"
+            if con is not None:
+                con.close()
+    
+        return result
 
     else:
         dados = request.json
 
-        conn = connectToDb()
-        cur = conn.cursor()
+        con = connectDb()
+        cur = con.cursor()
 
         try:
+            cur.execute('''select * from tb_carros where id = %s''', (id))
+            result = cur.fetchone()
+
+            if(result == None):
+                con.close()
+                return 'ERRO PUT\nCarro com id ' + id + ' não encontrado'
+            
             cur.execute('''
                         update tb_carros
                         set marca = %s, modelo = %s, ano = %s, cor = %s, valor = %s
@@ -68,50 +77,95 @@ def api_verbs():
                          dados['cor'],
                          dados['valor'],
                          dados['id']))
-            conn.commit()
+            con.commit()            
             cur.close()
         except (Exception, psycopg2.DatabaseError) as error:
             print(error)
         finally:
-            if conn is not None:
-                conn.close()
+            if con is not None:
+                con.close()
 
-        return("Carro atualizado com sucesso")
-
+        return result
 
 @carros_api.route('/<string:id>', methods=['GET'])
 def get_carros_by_id(id):
-    conn = connectToDb()
-    cur = conn.cursor(cursor_factory=RealDictCursor)
-
-    print(id)
+    con = connectDb()
+    cur = con.cursor(cursor_factory=RealDictCursor)
 
     try:
         cur.execute('''select * from tb_carros where id = %s''', (id))
         result = cur.fetchone()
         cur.close()
+
+        if(result == None):
+            return 'ERRO GET\nCarro com id ' + id + ' não encontrado'
+
     except (Exception, psycopg2.DatabaseError) as error:
         print(error)
     finally:
-        if conn is not None:
-            conn.close()
+        if con is not None:
+            con.close()
 
     return jsonify(result)
 
-@carros_api.route('/<string:id>', methods=['DELETE'])
-def delete(id):
-    conn = connectToDb()
-    cur = conn.cursor()
+@carros_api.route('/<string:id>', methods=['PUT'])
+def put(id):
+    dados = request.json
+
+    con = connectDb()
+    cur = con.cursor()
 
     try:
-        cur.execute('''delete from tb_carros where id = %s ''', (id,))
-        conn.commit()
+        cur.execute('''select * from tb_carros where id = %s''', (id))
+        result = cur.fetchone()
+
+        if(result == None):
+            con.close()
+            return 'ERRO PUT\nCarro com id ' + id + ' não encontrado'
+
+        cur.execute('''
+                    update tb_carros
+                    set marca = %s, modelo = %s, ano = %s, cor = %s, valor = %s
+                    where id = %s
+                    ''',
+                    (dados['marca'],
+                    dados['modelo'],
+                    dados['ano'],
+                    dados['cor'],
+                    dados['valor'],
+                    id))
+        con.commit()
+        result = get_carros_by_id(id)
         cur.close()
     except (Exception, psycopg2.DatabaseError) as error:
         print(error)
     finally:
-        if conn is not None:
-            conn.close()
+        if con is not None:
+            con.close()
 
-    return "Carro deletado com sucesso"
+    return result
+
+@carros_api.route('/<string:id>', methods=['DELETE'])
+def delete(id):
+    con = connectDb()
+    cur = con.cursor()
+
+    try:
+        cur.execute('''select * from tb_carros where id = %s''', (id))
+        result = cur.fetchone()
+
+        if(result == None):
+            con.close()
+            return 'ERRO DELETE\nCarro com id ' + id + ' não encontrado'
+        
+        cur.execute('''delete from tb_carros where id = %s ''', (id,))
+        con.commit()
+        cur.close()
+    except (Exception, psycopg2.DatabaseError) as error:
+        print(error)
+    finally:
+        if con is not None:
+            con.close()
+
+    return "Carro " + id + " deletado com sucesso"
